Rework entire game initialization sequence
To avoid the mess we have right now in the main application file.
