+++ VisualStudio/fheroes2/sources.props
@@ -137,6 +137,7 @@
     <ClCompile Include="src\fheroes2\game\game_exit.cpp" />
     <ClCompile Include="src\fheroes2\game\game_highscores.cpp" />
     <ClCompile Include="src\fheroes2\game\game_hotkeys.cpp" />
+    <ClCompile Include="src\fheroes2\game\game_init.cpp" />
     <ClCompile Include="src\fheroes2\game\game_interface.cpp" />
     <ClCompile Include="src\fheroes2\game\game_intro.cpp" />
     <ClCompile Include="src\fheroes2\game\game_invalid_assets.cpp" />
@@ -358,6 +359,7 @@
     <ClInclude Include="src\fheroes2\game\game_delays.h" />
     <ClInclude Include="src\fheroes2\game\game_exit.h" />
     <ClInclude Include="src\fheroes2\game\game_hotkeys.h" />
+    <ClInclude Include="src\fheroes2\game\game_init.h" />
     <ClInclude Include="src\fheroes2\game\game_interface.h" />
     <ClInclude Include="src\fheroes2\game\game_intro.h" />
     <ClInclude Include="src\fheroes2\game\game_invalid_assets.h" />
