+++ src/fheroes2/game/game_delays.h
@@ -102,4 +102,6 @@ namespace Game
 
     // Custom delay must never be called in this function.
     uint64_t getAnimationDelayValue( const DelayType delayType );
+
+    void initAnimateDelays();
 }
